refactor: Move user entities to domain/users package
- Reorganize User, Driver, Passenger, Admin, DriverDocument to domain/users package
- Update User entity to support large profile pictures (TEXT column)
- Update all imports across controllers, services, repositories, and entities
- Maintain JOINED inheritance strategy and all existing functionality
