--- drumigo/src/main/java/com/ftn/drumigo/domain/DriverDocument.java
@@ -1,36 +1,0 @@
-package com.ftn.drumigo.domain;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.Instant;
-
-@Entity
-@Table(name = "driver_documents")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class DriverDocument {
-    
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "driver_id", nullable = false)
-    private Driver driver;
-    
-    @Column(name = "document_name", nullable = false)
-    private String documentName;
-    
-    @Column(name = "document_url", length = 500)
-    private String documentUrl;
-    
-    @Column(name = "uploaded_at", nullable = false, updatable = false)
-    private Instant uploadedAt = Instant.now();
-}
-
