+++ drumigo/src/main/java/com/ftn/drumigo/domain/users/Admin.java
@@ -1,0 +1,15 @@
+package com.ftn.drumigo.domain.users;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Table(name = "admins")
+@PrimaryKeyJoinColumn(name = "user_id")
+@Getter
+@Setter
+public class Admin extends User {
+}
